# docs/sidecar/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RUDI Sidecar API",
    "version": "0.1.0",
    "description": "Machine-readable contract for the hardened RUDI sidecar surfaces: health, projects, notes, stable session endpoints, shell, terminal, filesystem, run-groups, and the public run-group WebSocket events.",
    "x-rudi-cli-version": "1.10.12"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:{port}",
      "description": "Local sidecar server",
      "variables": {
        "port": {
          "default": "8100",
          "description": "Dynamic sidecar port written to ~/.rudi/.rudi-lite-port"
        }
      }
    }
  ],
  "security": [
    {
      "RudiTokenAuth": []
    }
  ],
...
  "paths": {
...
    "/fs/read": {
      "get": {
        "tags": [
          "Filesystem"
        ],
        "summary": "Read a UTF-8 text file",
        "operationId": "readFileText",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AbsolutePath"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File contents",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FsReadResponse"
                },
                "example": {
                  "content": "hello world"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/fs/write": {
      "post": {
        "tags": [
          "Filesystem"
        ],
        "summary": "Write a UTF-8 text file",
        "description": "Creates parent directories automatically. The request body is capped at 50 MB.",
        "operationId": "writeFileText",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FsWriteRequest"
              },
              "example": {
                "path": "/Users/hoff/dev/RUDI/tmp/example.txt",
                "content": "hello world"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Write complete",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLargeError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/fs/write-binary": {
      "post": {
        "tags": [
          "Filesystem"
        ],
        "summary": "Write a binary file from base64 data",
        "description": "Creates parent directories automatically. The request body is capped at 50 MB.",
        "operationId": "writeFileBinary",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FsWriteBinaryRequest"
              },
              "example": {
                "path": "/Users/hoff/dev/RUDI/tmp/image.bin",
                "base64": "AAEC/w=="
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Binary write complete",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLargeError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/fs/readdir": {
      "get": {
        "tags": [
          "Filesystem"
        ],
        "summary": "List directory entries",
        "description": "Dotfiles are hidden by default. Results are cached briefly inside the sidecar.",
        "operationId": "readDirectory",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AbsolutePath"
            }
          },
          {
            "name": "showHidden",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            },
            "description": "Set to `1` to include dotfiles."
          }
        ],
        "responses": {
          "200": {
            "description": "Directory entries",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FsReaddirResponse"
                },
                "example": {
                  "entries": [
                    {
                      "name": "example.txt",
                      "path": "/Users/hoff/dev/RUDI/tmp/example.txt",
                      "isDirectory": false,
                      "isFile": true,
                      "size": 11,
                      "mtime": "2026-03-22T12:00:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/fs/stat": {
      "get": {
        "tags": [
          "Filesystem"
        ],
        "summary": "Read file or directory metadata",
        "operationId": "statFile",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AbsolutePath"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filesystem stat",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FsEntry"
                },
                "example": {
                  "name": "example.txt",
                  "path": "/Users/hoff/dev/RUDI/tmp/example.txt",
                  "isDirectory": false,
                  "isFile": true,
                  "size": 11,
                  "mtime": "2026-03-22T12:00:00.000Z"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/fs/serve": {
      "get": {
        "tags": [
          "Filesystem"
        ],
        "summary": "Serve a binary file",
        "description": "Streams a local file with a content type inferred from extension. The path must be an absolute local filesystem path.",
        "operationId": "serveFile",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AbsolutePath"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File stream",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "304": {
            "description": "Cached copy is current",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/fs/mkdir": {
      "post": {
        "tags": [
          "Filesystem"
        ],
        "summary": "Create a directory recursively",
        "operationId": "makeDirectory",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FsPathRequest"
              },
              "example": {
                "path": "/Users/hoff/dev/RUDI/tmp/nested"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Directory created",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/fs/remove": {
      "post": {
        "tags": [
          "Filesystem"
        ],
        "summary": "Remove a file or directory",
        "operationId": "removePath",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FsDestructivePathRequest"
              },
              "example": {
                "path": "/Users/hoff/dev/RUDI/tmp/example.txt",
                "confirmDestructive": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Path removed",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/fs/rename": {
      "post": {
        "tags": [
          "Filesystem"
        ],
        "summary": "Rename or move a file or directory",
        "operationId": "renamePath",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FsRenameRequest"
              },
              "example": {
                "oldPath": "/Users/hoff/dev/RUDI/tmp/example.txt",
                "newPath": "/Users/hoff/dev/RUDI/tmp/example-renamed.txt"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Path renamed",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/fs/watch": {
      "post": {
        "tags": [
          "Filesystem"
        ],
        "summary": "Watch a filesystem path for sidecar change events",
        "description": "Registers an in-process filesystem watcher. The path must be absolute and cannot be the filesystem root.",
        "operationId": "watchPath",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FsPathRequest"
              },
              "example": {
                "path": "/Users/hoff/dev/RUDI/tmp"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Watch registered",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/fs/unwatch": {
      "post": {
        "tags": [
          "Filesystem"
        ],
        "summary": "Stop watching a filesystem path",
        "description": "Unregisters an in-process filesystem watcher. The path must be absolute and cannot be the filesystem root.",
        "operationId": "unwatchPath",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FsPathRequest"
              },
              "example": {
                "path": "/Users/hoff/dev/RUDI/tmp"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Watch removed",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "RudiTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-rudi-token",
        "description": "Sidecar auth token read from ~/.rudi/.rudi-lite-token."
      }
    },
    "headers": {
      "RequestIdHeader": {
        "description": "Per-request correlation ID returned on sidecar responses.",
        "schema": {
          "type": "string"
        }
      }
    },
...
    "responses": {
      "UnauthorizedError": {
        "description": "Unauthorized",
        "headers": {
          "x-rudi-request-id": {
            "$ref": "#/components/headers/RequestIdHeader"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SidecarError"
            },
            "example": {
              "error": "Unauthorized",
              "code": "UNAUTHORIZED",
              "requestId": "req_example_123"
            }
          }
        }
      },
...
      "NotFoundError": {
        "description": "Not found",
        "headers": {
          "x-rudi-request-id": {
            "$ref": "#/components/headers/RequestIdHeader"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SidecarError"
            },
            "example": {
              "error": "Not found",
              "code": "NOT_FOUND",
              "requestId": "req_example_123"
            }
          }
        }
      },
      "RequestTooLargeError": {
        "description": "Request body too large",
        "headers": {
          "x-rudi-request-id": {
            "$ref": "#/components/headers/RequestIdHeader"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SidecarError"
            },
            "example": {
              "error": "Request body too large",
              "code": "REQUEST_TOO_LARGE",
              "requestId": "req_example_123"
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Missing required field or invalid field value.",
        "headers": {
          "x-rudi-request-id": {
            "$ref": "#/components/headers/RequestIdHeader"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SidecarError"
            },
            "examples": {
              "missingRequiredField": {
                "value": {
                  "error": "path required",
                  "code": "MISSING_REQUIRED_FIELD",
                  "details": {
                    "field": "path",
                    "location": "body"
                  },
                  "requestId": "req_example_123"
                }
              },
              "invalidPath": {
                "value": {
                  "error": "path must be an absolute filesystem path",
                  "code": "INVALID_FIELD",
                  "details": {
                    "field": "path",
                    "location": "body",
                    "reason": "absolute_path_required"
                  },
                  "requestId": "req_example_123"
                }
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal server error",
        "headers": {
          "x-rudi-request-id": {
            "$ref": "#/components/headers/RequestIdHeader"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SidecarError"
            },
            "example": {
              "error": "Internal server error",
              "code": "INTERNAL_ERROR",
              "requestId": "req_example_123"
            }
          }
        }
      }
    },
    "schemas": {
...
      "OkResponse": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          }
        }
      },
...
      "AbsolutePath": {
        "type": "string",
        "description": "Absolute local filesystem path. Empty, relative, and NUL-containing values are rejected.",
        "examples": [
          "/Users/hoff/dev/RUDI/tmp/example.txt"
        ]
      },
...
      "FsReadResponse": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "FsReaddirResponse": {
        "type": "object",
        "required": [
          "entries"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FsEntry"
            }
          }
        },
        "additionalProperties": false
      },
...
      "FsWriteRequest": {
        "type": "object",
        "required": [
          "path",
          "content"
        ],
        "properties": {
          "path": {
            "$ref": "#/components/schemas/MutableAbsolutePath"
          },
          "content": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "FsWriteBinaryRequest": {
        "type": "object",
        "required": [
          "path",
          "base64"
        ],
        "properties": {
          "path": {
            "$ref": "#/components/schemas/MutableAbsolutePath"
          },
          "base64": {
            "type": "string",
            "pattern": "^(?:[A-Za-z0-9+/]{4})*(?:[A-Za-z0-9+/]{2}==|[A-Za-z0-9+/]{3}=)?$",
            "description": "Strict standard base64 content without whitespace or data URI prefixes."
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}