# src/commands/serve/routes/fs.js
/**
 * File system operations — read, write, readdir, stat, serve, watch/unwatch.
 *
 * Owns: fsWatchers Map, fsReaddirCache Map, fsReaddirInFlight Map, generation counter.
 */

import fsSync from 'fs';
import fsp from 'fs/promises';
import path from 'path';
import { rejectInvalidPathField, rejectMissingDestructiveConfirmation } from '../validation.js';

const FS_READDIR_CACHE_TTL_MS = 1200;
const MAX_FS_WRITE_BODY_SIZE = 50 * 1024 * 1024;
const BASE64_PATTERN = /^(?:[A-Za-z0-9+/]{4})*(?:[A-Za-z0-9+/]{2}==|[A-Za-z0-9+/]{3}=)?$/;

export function buildFsRoutes(ctx) {
  const { json, error, readBody, log, broadcast, requiredField, requiredFields, invalidField } = ctx;

  const fsWatchers = new Map(); // path -> { watcher, debounceTimer }
  const fsReaddirCache = new Map(); // key -> { entries, fetchedAt }
  const fsReaddirInFlight = new Map(); // key -> Promise<entries>
  let fsReaddirCacheGeneration = 0;

  function invalidateFsReaddirCache() {
    fsReaddirCacheGeneration += 1;
    fsReaddirCache.clear();
  }

  function getFsReaddirCacheKey(dirPath, showHidden) {
    return `${showHidden ? '1' : '0'}:${dirPath}`;
  }

  function rejectFsPath(value, res, options = {}) {
    return rejectInvalidPathField({
      value,
      res,
      invalidField,
      error,
      ...options,
    });
  }

  function rejectInvalidBase64(value, res) {
    if (typeof value !== 'string' || !BASE64_PATTERN.test(value)) {
      invalidField(res, 'base64', 'base64 must be a valid base64 string', {
        reason: typeof value === 'string' ? 'invalid_base64' : 'invalid_type',
      });
      return true;
    }

    return false;
  }

  async function readDirectoryEntries(dirPath, showHidden) {
    const cacheKey = getFsReaddirCacheKey(dirPath, showHidden);
    const now = Date.now();
    const cached = fsReaddirCache.get(cacheKey);
    if (cached && (now - cached.fetchedAt) <= FS_READDIR_CACHE_TTL_MS) {
      return cached.entries;
    }

    const inFlight = fsReaddirInFlight.get(cacheKey);
    if (inFlight) {
      return inFlight;
    }

    const generationAtStart = fsReaddirCacheGeneration;
    const request = (async () => {
      const names = await fsp.readdir(dirPath);
      const entries = await Promise.all(
        names
          .filter(n => showHidden || !n.startsWith('.'))
          .map(async (name) => {
            const fullPath = path.join(dirPath, name);
            try {
              const stat = await fsp.stat(fullPath);
              return {
                name,
                path: fullPath,
                isDirectory: stat.isDirectory(),
                isFile: stat.isFile(),
                size: stat.size,
                mtime: stat.mtime.toISOString(),
              };
            } catch {
              return null;
            }
          }),
      );
      return entries.filter(Boolean);
    })();

    fsReaddirInFlight.set(cacheKey, request);
    try {
      const entries = await request;
      if (generationAtStart === fsReaddirCacheGeneration) {
        fsReaddirCache.set(cacheKey, { entries, fetchedAt: Date.now() });
      }
      return entries;
    } finally {
      fsReaddirInFlight.delete(cacheKey);
    }
  }

  async function handle(req, res, url) {
    const pathname = url.pathname;

    // GET /fs/read?path=
    if (req.method === 'GET' && pathname === '/fs/read') {
      const filePath = url.searchParams.get('path');
      if (!filePath) return requiredField(res, 'path', { location: 'query' });
      if (rejectFsPath(filePath, res, { field: 'path', location: 'query' })) return true;
      try {
        const content = await fsp.readFile(filePath, 'utf-8');
        json(res, { content });
      } catch (err) {
        error(res, err.message, 404);
      }
      return true;
    }

    // POST /fs/write {path, content}
    if (req.method === 'POST' && pathname === '/fs/write') {
      const body = await readBody(req, { maxBodySize: MAX_FS_WRITE_BODY_SIZE });
      if (!body.path || body.content === undefined) {
        const missing = [];
        if (!body.path) missing.push('path');
        if (body.content === undefined) missing.push('content');
        return requiredFields(res, missing);
      }
      if (rejectFsPath(body.path, res, { allowRoot: false })) return true;
      try {
        await fsp.mkdir(path.dirname(body.path), { recursive: true });
        await fsp.writeFile(body.path, body.content, 'utf-8');
        invalidateFsReaddirCache();
        json(res, { ok: true });
      } catch (err) {
        error(res, err.message, 500);
      }
      return true;
    }

    // POST /fs/write-binary {path, base64}
    if (req.method === 'POST' && pathname === '/fs/write-binary') {
      const body = await readBody(req, { maxBodySize: MAX_FS_WRITE_BODY_SIZE });
      if (!body.path || body.base64 === undefined) {
        const missing = [];
        if (!body.path) missing.push('path');
        if (body.base64 === undefined) missing.push('base64');
        return requiredFields(res, missing);
      }
      if (rejectFsPath(body.path, res, { allowRoot: false })) return true;
      if (rejectInvalidBase64(body.base64, res)) return true;
      try {
        await fsp.mkdir(path.dirname(body.path), { recursive: true });
        const buffer = Buffer.from(body.base64, 'base64');
        await fsp.writeFile(body.path, buffer);
        invalidateFsReaddirCache();
        json(res, { ok: true });
      } catch (err) {
        error(res, err.message, 500);
      }
      return true;
    }

    // GET /fs/readdir?path=&showHidden=1
    if (req.method === 'GET' && pathname === '/fs/readdir') {
      const dirPath = url.searchParams.get('path');
      if (!dirPath) return requiredField(res, 'path', { location: 'query' });
      if (rejectFsPath(dirPath, res, { field: 'path', location: 'query' })) return true;
      const showHidden = url.searchParams.get('showHidden') === '1';
      try {
        const entries = await readDirectoryEntries(dirPath, showHidden);
        json(res, { entries });
      } catch (err) {
        error(res, err.message, 404);
      }
      return true;
    }

    // GET /fs/stat?path=
    if (req.method === 'GET' && pathname === '/fs/stat') {
      const filePath = url.searchParams.get('path');
      if (!filePath) return requiredField(res, 'path', { location: 'query' });
      if (rejectFsPath(filePath, res, { field: 'path', location: 'query' })) return true;
      try {
        const stat = await fsp.stat(filePath);
        json(res, {
          name: path.basename(filePath),
          path: filePath,
          isDirectory: stat.isDirectory(),
          isFile: stat.isFile(),
          size: stat.size,
          mtime: stat.mtime.toISOString(),
        });
      } catch (err) {
        error(res, err.message, 404);
      }
      return true;
    }

    // GET /fs/serve?path= (binary file serving)
    if (req.method === 'GET' && pathname === '/fs/serve') {
      const filePath = url.searchParams.get('path');
      if (!filePath) return requiredField(res, 'path', { location: 'query' });
      if (rejectFsPath(filePath, res, { field: 'path', location: 'query' })) return true;
      try {
        const stat = await fsp.stat(filePath);
        const ext = path.extname(filePath).toLowerCase();
        const mimeTypes = {
          '.png': 'image/png', '.jpg': 'image/jpeg', '.jpeg': 'image/jpeg',
          '.gif': 'image/gif', '.svg': 'image/svg+xml', '.webp': 'image/webp',
          '.pdf': 'application/pdf', '.mp4': 'video/mp4', '.webm': 'video/webm',
          '.mp3': 'audio/mpeg', '.wav': 'audio/wav',
          '.json': 'application/json', '.csv': 'text/csv',
          '.html': 'text/html', '.txt': 'text/plain',
        };
        const contentType = mimeTypes[ext] || 'application/octet-stream';
        const etag = `"${stat.mtimeMs.toString(36)}-${stat.size.toString(36)}"`;
        if (req.headers['if-none-match'] === etag) {
          res.writeHead(304, { 'Access-Control-Allow-Origin': '*' });
          res.end();
          return true;
        }
        res.writeHead(200, {
          'Content-Type': contentType,
          'Content-Length': stat.size,
          'Access-Control-Allow-Origin': '*',
          'Cache-Control': 'public, max-age=5',
          'ETag': etag,
        });
        fsSync.createReadStream(filePath).pipe(res);
      } catch (err) {
        error(res, err.message, 404);
      }
      return true;
    }

    // POST /fs/mkdir {path}
    if (req.method === 'POST' && pathname === '/fs/mkdir') {
      const body = await readBody(req);
      if (!body.path) return requiredField(res, 'path');
      if (rejectFsPath(body.path, res, { allowRoot: false })) return true;
      try {
        await fsp.mkdir(body.path, { recursive: true });
        invalidateFsReaddirCache();
        json(res, { ok: true });
      } catch (err) {
        error(res, err.message, 500);
      }
      return true;
    }

    // POST /fs/remove {path}
    if (req.method === 'POST' && pathname === '/fs/remove') {
      const body = await readBody(req);
      if (!body.path) return requiredField(res, 'path');
      if (rejectFsPath(body.path, res, { allowRoot: false })) return true;
      if (rejectMissingDestructiveConfirmation({ body, res, invalidField, error, operation: 'fs remove' })) {
        return true;
      }
      try {
        await fsp.rm(body.path, { recursive: true });
        invalidateFsReaddirCache();
        json(res, { ok: true });
      } catch (err) {
        error(res, err.message, 500);
      }
      return true;
    }

    // POST /fs/rename {oldPath, newPath}
    if (req.method === 'POST' && pathname === '/fs/rename') {
      const body = await readBody(req);
      if (!body.oldPath || !body.newPath) {
        const missing = [];
        if (!body.oldPath) missing.push('oldPath');
        if (!body.newPath) missing.push('newPath');
        return requiredFields(res, missing);
      }
      if (rejectFsPath(body.oldPath, res, { field: 'oldPath', allowRoot: false })) return true;
      if (rejectFsPath(body.newPath, res, { field: 'newPath', allowRoot: false })) return true;
      try {
        await fsp.rename(body.oldPath, body.newPath);
        invalidateFsReaddirCache();
        json(res, { ok: true });
      } catch (err) {
        error(res, err.message, 500);
      }
      return true;
    }

    // POST /fs/watch {path}
    if (req.method === 'POST' && pathname === '/fs/watch') {
      const body = await readBody(req);
      if (!body.path) return requiredField(res, 'path');
      if (rejectFsPath(body.path, res, { allowRoot: false })) return true;
      const watchPath = body.path;

      if (fsWatchers.has(watchPath)) {
        json(res, { ok: true, already: true });
        return true;
      }

      try {
        const watcher = fsSync.watch(watchPath, { recursive: true }, (eventType, filename) => {
          if (!filename) return;
          const entry = fsWatchers.get(watchPath);
          if (!entry) return;

          clearTimeout(entry.debounceTimer);
          entry.debounceTimer = setTimeout(() => {
            const fullPath = path.join(watchPath, filename);
            const dirPath = path.dirname(fullPath);
            invalidateFsReaddirCache();
            broadcast('fs:change', { event: eventType, path: fullPath, dir: dirPath });
          }, 100);
        });

        fsWatchers.set(watchPath, { watcher, debounceTimer: null });
        log('fs', 'info', 'watching filesystem path');
        json(res, { ok: true });
      } catch (err) {
        error(res, err.message, 500);
      }
      return true;
    }

    // POST /fs/unwatch {path}
    if (req.method === 'POST' && pathname === '/fs/unwatch') {
      const body = await readBody(req);
      if (!body.path) return requiredField(res, 'path');
      if (rejectFsPath(body.path, res, { allowRoot: false })) return true;
      const entry = fsWatchers.get(body.path);
      if (entry) {
        clearTimeout(entry.debounceTimer);
        entry.watcher.close();
        fsWatchers.delete(body.path);
        log('fs', 'info', 'unwatched filesystem path');
      }
      json(res, { ok: true });
      return true;
    }

    return false;
  }

  function cleanup() {
    for (const [, entry] of fsWatchers) {
      try {
        clearTimeout(entry.debounceTimer);
        entry.watcher.close();
      } catch (err) {
        log('fs', 'warn', `failed to close filesystem watcher during cleanup: ${err.message}`);
      }
    }
    fsWatchers.clear();
  }

  return { handle, cleanup };
}